Move rule-engine folder to root of src/
