+++ packages/@coorpacademy-progression-engine/src/reducers/variables.js
@@ -1,7 +1,7 @@
 // @flow
 
 import type {Action, State, Config} from '../types';
-import updateVariables from '../compute-next-step/rule-engine/apply-instructions';
+import updateVariables from '../rule-engine/apply-instructions';
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default (config: Config) => (state: State, action: Action): State => {
