+++ packages/@coorpacademy-progression-engine/src/rule-engine/test/select-rule.js
